requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
arrow==0.15.7
asn1crypto==0.24.0
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
binaryornot==0.4.4
black==19.10b0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
conda==4.7.12
conda-package-handling==1.5.0
cookiecutter==1.7.2
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
dnspython==1.16.0
docxtpl==0.10.0
entrypoints==0.3
hupper==1.10.2
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
jinja2-time==0.2.0
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.4.0
jupyterlab==1.1.3
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Mako==1.1.3
markdown2==2.3.9
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
parso==0.5.1
PasteDeploy==2.1.0
pathspec==0.8.0
pickleshare==0.7.5
plaster==1.0
plaster-pastedeploy==0.7
poyo==0.5.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pylint==2.5.3
pymongo==3.10.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyramid==1.10.4
pyramid-debugtoolbar==4.6.1
pyramid-jinja2==2.8
pyramid-mako==1.1.0
pyrsistent==0.15.4
PySocks==1.7.0
python-dateutil==2.8.0
python-docx==0.8.10
python-slugify==4.0.1
pytz==2020.1
pywin32==223
pywinpty==0.5.5
pyzmq==18.1.0
regex==2020.6.8
repoze.lru==0.7
requests==2.24.0
ruamel-yaml==0.15.46
secretary==0.2.19
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
toml==0.10.1
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
translationstring==1.3
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.24.2
venusian==3.0.0
waitress==1.4.4
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.6
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zope.deprecation==4.4.0
zope.interface==5.1.0
